fix report progress bug

+++ internal/transceiver/util.go
@@ -2,14 +2,14 @@ package transceiver
 
 import "log"
 
-func reportProgress(startTimeNs int64, nowNs int64, sentMsg uint64) uint32 {
-	elapsedSec := uint32((nowNs - startTimeNs)) / NANOS_PER_SECOND
-	var sendRate uint32
+func reportProgress(startTimeNs int64, nowNs int64, sentMsg uint64) uint64 {
+	elapsedSec := uint64((nowNs - startTimeNs)) / NANOS_PER_SECOND
+	var sendRate uint64
 	switch elapsedSec {
 	case 0:
-		sendRate = uint32(sentMsg)
+		sendRate = sentMsg
 	default:
-		sendRate = uint32(sentMsg) / elapsedSec
+		sendRate = sentMsg / elapsedSec
 	}
 	log.Printf("[info] send rate %d msg/sec\n", sendRate)
 	return elapsedSec
